@@ -22,51 +22,52 @@
         "bundle-report": "webpack-bundle-analyzer dist/stats.json"
     },
     "dependencies": {
-        "@angular/animations": "^7.1.4",
-        "@angular/cdk": "^7.2.0",
-        "@angular/common": "^7.1.4",
-        "@angular/compiler": "^7.1.4",
-        "@angular/core": "^7.1.4",
-        "@angular/flex-layout": "^7.0.0-beta.22",
-        "@angular/forms": "^7.1.4",
-        "@angular/http": "^7.1.4",
-        "@angular/material": "^7.2.0",
-        "@angular/material-moment-adapter": "^7.2.0",
-        "@angular/platform-browser": "^7.1.4",
-        "@angular/platform-browser-dynamic": "^7.1.4",
-        "@angular/router": "^7.1.4",
+        "@angular/animations": "^7.2.0",
+        "@angular/cdk": "^7.2.1",
+        "@angular/common": "^7.2.0",
+        "@angular/compiler": "^7.2.0",
+        "@angular/core": "^7.2.0",
+        "@angular/flex-layout": "^7.0.0-beta.23",
+        "@angular/forms": "^7.2.0",
+        "@angular/http": "^7.2.0",
+        "@angular/material": "^7.2.1",
+        "@angular/material-moment-adapter": "^7.2.1",
+        "@angular/platform-browser": "^7.2.0",
+        "@angular/platform-browser-dynamic": "^7.2.0",
+        "@angular/router": "^7.2.0",
+        "@ckeditor/ckeditor5-angular": "^1.0.1",
+        "@ckeditor/ckeditor5-build-balloon": "^11.2.0",
+        "@ckeditor/ckeditor5-editor-balloon": "^11.0.2",
         "angular-oauth2-oidc": "5.0.2",
-        "angular-tree-component": "^8.0.4",
-        "ckeditor": "^4.11.1",
+        "angular-tree-component": "^8.2.0",
         "class-transformer": "^0.2.0",
         "classlist.js": "^1.1.20150312",
-        "core-js": "^2.6.1",
+        "core-js": "^2.6.2",
         "hammerjs": "^2.0.8",
         "moment": "2.23.0",
-        "ng2-ckeditor": "^1.2.2",
         "ngx-custom-validators": "^7.0.0",
-        "ngx-input-file": "^7.0.0",
+        "ngx-input-file": "^7.0.1",
         "ngx-moment": "3.3.0",
         "object-path": "^0.11.4",
         "perfect-scrollbar": "^1.4.0",
         "rxjs": "^6.3.3",
         "rxjs-compat": "^6.3.3",
         "web-animations-js": "^2.3.1",
-        "zone.js": "^0.8.26"
+        "zone.js": "^0.8.27"
     },
     "devDependencies": {
-        "@angular-devkit/build-angular": "^0.11.4",
-        "@angular/cli": "^7.1.4",
-        "@angular/compiler-cli": "^7.1.4",
-        "@angular/language-service": "^7.1.4",
+        "@angular-devkit/build-angular": "^0.12.1",
+        "@angular/cli": "^7.2.1",
+        "@angular/compiler-cli": "^7.2.0",
+        "@angular/language-service": "^7.2.0",
         "@angularclass/hmr": "^2.1.3",
         "@types/node": "^10.12.18",
-        "@types/object-path": "0.9.29",
+        "@types/object-path": "0.11.0",
         "codelyzer": "^4.5.0",
         "node-sass": "^4.11.0",
         "ts-node": "^7.0.1",
-        "tslint": "^5.12.0",
-        "typescript": "^3.1.6",
+        "tslint": "^5.12.1",
+        "typescript": "^3.2.2",
         "webpack-bundle-analyzer": "^3.0.3"
     }
 }
