@@ -1,65 +1,94 @@
-import {RestClient} from './HttpClient';
-import {ListResult} from './list/ListResult';
-import {plainToClass} from 'class-transformer';
-import {map} from 'rxjs/operators';
-import {SaveModelResponse} from './SaveModelResponse';
-import {ClassType} from 'class-transformer/ClassTransformer';
-import {StorageItem} from '../@models/results/StorageItem';
-import {Observable} from 'rxjs';
-import {Filter} from './Filter';
-import {InputFile} from 'ngx-input-file';
+import { RestClient } from './HttpClient';
+import { ListResult } from './list/ListResult';
+import { plainToClass } from 'class-transformer';
+import { map } from 'rxjs/operators';
+import { SaveModelResponse } from './SaveModelResponse';
+import { ClassType } from 'class-transformer/ClassTransformer';
+import { StorageItem } from '../@models/results/StorageItem';
+import { Observable } from 'rxjs';
+import { Filter } from './Filter';
+import { InputFile } from 'ngx-input-file';
 
 export abstract class BaseService<T> {
+    protected constructor(protected httpClient: RestClient) {}
 
-    protected constructor(protected httpClient: RestClient) {
-    }
-
-    public getAll(page: number = 0, perPage: number = 10, sort = '', filter: Filter = null): Observable<ListResult<T>> {
-        return this.httpClient.get(this.getResource(), {
-            limit: perPage || 10,
-            offset: page > 0 ? perPage * (page - 1) : 0,
-            order: sort,
-            filter: filter == null ? null : filter.toString()
-        }).pipe(map(res => plainToClass(this.getListType(), res as ListResult<T>)));
+    public getAll(
+        page: number = 0,
+        perPage: number = 10,
+        sort = '',
+        filter: Filter = null
+    ): Observable<ListResult<T>> {
+        return this.httpClient
+            .get(this.getResource(), {
+                limit: perPage || 10,
+                offset: page > 0 ? perPage * (page - 1) : 0,
+                order: sort,
+                filter: filter == null ? null : filter.toString()
+            })
+            .pipe(
+                map(res =>
+                    plainToClass(this.getListType(), res as ListResult<T>)
+                )
+            );
     }
 
     public get(id: number): Observable<T> {
-        return this.httpClient.get(this.getResource() + '/' + id, {})
+        return this.httpClient
+            .get(this.getResource() + '/' + id, {})
             .pipe(map(res => plainToClass(this.getType(), res as T)));
     }
 
     public new(): Observable<T> {
-        return this.httpClient.get(this.getResource() + '/new', {})
+        return this.httpClient
+            .get(this.getResource() + '/new', {})
             .pipe(map(res => plainToClass(this.getType(), res as T)));
     }
 
     public add(item: T): Observable<SaveModelResponse<T>> {
-        return this.httpClient.post(this.getResource(), item)
-            .pipe(map(res => plainToClass(this.getSaveType(), res as SaveModelResponse<T>)));
+        return this.httpClient
+            .post(this.getResource(), item)
+            .pipe(
+                map(res =>
+                    plainToClass(this.getSaveType(), res as SaveModelResponse<
+                        T
+                    >)
+                )
+            );
     }
 
     public update(id: number, item: T): Observable<SaveModelResponse<T>> {
-        return this.httpClient.put(this.getResource() + '/' + id, item)
-            .pipe(map(res => plainToClass(this.getSaveType(), res as SaveModelResponse<T>)));
+        return this.httpClient
+            .put(this.getResource() + '/' + id, item)
+            .pipe(
+                map(res =>
+                    plainToClass(this.getSaveType(), res as SaveModelResponse<
+                        T
+                    >)
+                )
+            );
     }
 
     public publish(id: number): Observable<T> {
-        return this.httpClient.post(this.getResource() + '/publish/' + id, {})
+        return this.httpClient
+            .post(this.getResource() + '/publish/' + id, {})
             .pipe(map(res => plainToClass(this.getType(), res as T)));
     }
 
     public unpublish(id: number): Observable<T> {
-        return this.httpClient.post(this.getResource() + '/hide/' + id, {})
+        return this.httpClient
+            .post(this.getResource() + '/hide/' + id, {})
             .pipe(map(res => plainToClass(this.getType(), res as T)));
     }
 
     public delete(id: number): Observable<boolean> {
-        return this.httpClient.delete(this.getResource() + '/' + id)
+        return this.httpClient
+            .delete(this.getResource() + '/' + id)
             .pipe(map(() => true));
     }
 
     public count(): Observable<number> {
-        return this.httpClient.get(this.getResource() + '/count', {})
+        return this.httpClient
+            .get(this.getResource() + '/count', {})
             .pipe(map(res => res as number));
     }
 
@@ -73,16 +102,18 @@ export abstract class BaseService<T> {
 }
 
 export interface IBaseServiceWithUpload {
-    upload(file: InputFile): Observable<StorageItem>;
+    upload(file: File): Observable<StorageItem>;
 }
 
 export interface IBaseServiceCreatable<T> {
     create(name: string): Observable<SaveModelResponse<T>>;
 }
 
-export abstract class BaseServiceWithUpload<T> extends BaseService<T> implements IBaseServiceWithUpload {
-    public upload(file: InputFile): Observable<StorageItem> {
-        return this.httpClient.post(this.getResource() + '/upload/', file.file, {name: file.file.name})
+export abstract class BaseServiceWithUpload<T> extends BaseService<T>
+    implements IBaseServiceWithUpload {
+    public upload(file: File): Observable<StorageItem> {
+        return this.httpClient
+            .post(this.getResource() + '/upload/', file, { name: file.name })
             .pipe(map(data => plainToClass(StorageItem, data as StorageItem)));
     }
 }
